#6239 : Mail Scanner Settings Not Functioning - Appu =====================================================
When setting up Mail Scanner, the settings will not take and one is redirected to a blank index.php page. If the settings are put into the database by hand, they are displayed properly in the WebUI, but are not editable. Also, the scanner does not appear to be able to scan for the correct folders in the given email account, but rules are addable and removable. -> Addressed this issue