#7182 : Tax name duplicate validation not present on edit - Asha =================================================================
Settings > Tax Calculations Trying to add a new Tax with the same name as an existing tax gives a message indicating that names cannot be repeated and the action is not taken. But if you edit the taxes and give two the same name, the save is done correctly and we get two (or more) taxes with the same name.